airbnb-backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
Pillow==10.0.0
stripe==5.0.0
pytest==7.2.2
pytest-django==4.5.2